MIM-24 收口自适应子会话导航
