+++ ios/MimiRemote/Tests/MimiRemoteTests/ConversationDirectRuntimeHistoryTests.swift
@@ -2010,72 +2010,6 @@ extension ConversationDataFlowTests {
         XCTAssertEqual(context.subagents.first?.parentThreadID, "thr_parent")
     }
 
-    func testSessionContextStoreMergesUpdatesAndAttachesSubagents() {
-        let store = SessionContextStore()
-        store.upsert(
-            SessionContextSnapshot(
-                sessionID: "thr_parent",
-                threadID: "thr_parent",
-                status: SessionContextStatus(type: "idle"),
-                environment: SessionContextEnvironment(id: "local", kind: "local", label: "本地", cwd: "/tmp/parent", provider: "openai"),
-                sources: [SessionContextSource(id: "session_source", kind: "session", label: "appServer")]
-            ),
-            fallbackSessionID: nil
-        )
-        store.upsert(
-            SessionContextSnapshot(
-                sessionID: "thr_parent",
-                status: SessionContextStatus(type: "active", activeFlags: ["waitingOnApproval"]),
-                tasks: [SessionContextTask(id: "cmd_1", kind: "command", title: "go test ./...", subtitle: "/tmp/parent", status: "running")]
-            ),
-            fallbackSessionID: nil
-        )
-        store.upsert(
-            SessionContextSnapshot(
-                sessionID: "thr_child",
-                threadID: "thr_child",
-                subagents: [
-                    SessionContextSubagent(
-                        id: "thr_child",
-                        parentThreadID: "thr_parent",
-                        sessionID: "session_child",
-                        nickname: "Noether",
-                        role: "review",
-                        status: "running",
-                        canAcceptDirectInput: false
-                    )
-                ]
-            ),
-            fallbackSessionID: nil
-        )
-        store.upsert(
-            SessionContextSnapshot(
-                sessionID: "thr_child",
-                threadID: "thr_child",
-                subagents: [
-                    SessionContextSubagent(
-                        id: "thr_child",
-                        status: "completed",
-                        statusMessage: "done",
-                        canAcceptDirectInput: true
-                    )
-                ]
-            ),
-            fallbackSessionID: nil
-        )
-
-        let parent = store.context(for: "thr_parent")
-        XCTAssertEqual(parent?.status?.activeFlags, ["waitingOnApproval"])
-        XCTAssertEqual(parent?.environment?.cwd, "/tmp/parent")
-        XCTAssertEqual(parent?.tasks.first?.title, "go test ./...")
-        XCTAssertEqual(parent?.subagents.first?.displayName, "Noether")
-        XCTAssertEqual(parent?.subagents.first?.sessionID, "session_child")
-        XCTAssertEqual(parent?.subagents.first?.status, "completed")
-        XCTAssertEqual(parent?.subagents.first?.statusMessage, "done")
-        XCTAssertEqual(parent?.subagents.first?.canAcceptDirectInput, false)
-        XCTAssertEqual(store.context(for: "codex_thr_parent")?.subagents.first?.id, "thr_child")
-    }
-
     func testSessionContextStoreClearsPendingApprovalTasks() {
         let store = SessionContextStore()
         store.upsert(
