+++ ios/MimiRemote/Tests/MimiRemoteTests/ThemeStoreTests.swift
@@ -482,6 +482,7 @@ final class ResponsiveLayoutTests: XCTestCase {
         ))
         XCTAssertTrue(layout.prefersDetailOnly)
         XCTAssertFalse(layout.usesAttachedInspector)
+        XCTAssertFalse(layout.usesSheetInspectorNavigation)
         XCTAssertLessThanOrEqual(layout.titleMaxWidth, 230)
         XCTAssertGreaterThanOrEqual(layout.titleMaxWidth, 160)
     }
@@ -495,6 +496,16 @@ final class ResponsiveLayoutTests: XCTestCase {
         ))
         XCTAssertTrue(layout.prefersDetailOnly)
         XCTAssertFalse(layout.usesAttachedInspector)
+        XCTAssertFalse(layout.usesSheetInspectorNavigation)
+    }
+
+    func testWorkbenchLayoutUsesSheetNavigationOnMediumIPadWidth() {
+        let layout = WorkbenchLayout(containerWidth: 1_032, horizontalSizeClass: .regular)
+
+        XCTAssertFalse(layout.usesCompactNavigation)
+        XCTAssertFalse(layout.prefersDetailOnly)
+        XCTAssertFalse(layout.usesAttachedInspector)
+        XCTAssertTrue(layout.usesSheetInspectorNavigation)
     }
 
     func testWorkbenchLayoutKeepsSplitNavigationOnWidePadWidth() {
@@ -506,6 +517,7 @@ final class ResponsiveLayoutTests: XCTestCase {
         ))
         XCTAssertFalse(layout.prefersDetailOnly)
         XCTAssertTrue(layout.usesAttachedInspector)
+        XCTAssertFalse(layout.usesSheetInspectorNavigation)
         XCTAssertEqual(layout.projectColumn.ideal, 330)
         XCTAssertEqual(layout.titleMaxWidth, 340)
     }
